requirements.txt
email-validator>=2.2.0
flask-socketio>=5.5.1
flask>=3.1.0
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
psycopg2-binary>=2.9.10
cryptography>=44.0.1
werkzeug>=3.1.3
flask-login>=0.6.3
flask-wtf>=1.2.2
argon2-cffi>=23.1.0
setuptools>=75.8.1
websockets>=15.0
pynacl>=1.5.0
requests[socks]>=2.31.0
pycryptodome>=3.19.0
pyi2p>=1.3.0
flask-limiter>=3.5.0
python-dotenv>=1.0.1
